linter: improvements
- Improved message for argCount checker;

- Fixed PHPDoc handling for closures and also fixed return
  type deduction so it now takes PHPDoc into account;

- Type tests for closures are moved to a separate file;

- Types for shapes are now displayed as shape{args...}.
