+++ src/linter/autogen/autogen.go
@@ -41,3 +41,24 @@ func GenerateClosureName(fun *ir.ClosureExpr, currentFunction, currentFile strin
 	}
 	return fmt.Sprintf("\\Closure$(%s%s):%d$", currentFile, curFunction, pos.StartLine)
 }
+
+func TransformClosureToReadableName(name string) string {
+	name = strings.TrimSuffix(name, "$")
+	if types.IsClosureFromPHPDoc(name) {
+		name = strings.TrimPrefix(name, `\Closure$(`)
+		parts := strings.Split(name, ")")
+		args := parts[0]
+		returnType := strings.ReplaceAll(parts[1], "/", "|")
+		returnType = strings.TrimPrefix(returnType, ":")
+		if returnType == "" {
+			returnType = "void"
+		}
+		return fmt.Sprintf("anonymous(%s): %s defined in PHPDoc", args, returnType)
+	}
+
+	index := strings.LastIndexByte(name, ':')
+	if index == -1 {
+		return name
+	}
+	return fmt.Sprintf("anonymous(...) defined on line %s", name[index+1:])
+}
